refactoring test location
